Move FlagResetRule from core to api/testFixtures
