+++ api/src/testFixtures/java/io/grpc/FlagResetRule.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.grpc.internal;
+package io.grpc;
 
 import java.util.ArrayDeque;
 import java.util.Deque;
